Optimization

+++ src/main2.py
@@ -151,7 +151,7 @@ def main():
     p.add_argument("--resume", type=str, default=None)
     p.add_argument("--save_freq", type=int, default=3)
     p.add_argument("--mixed_precision", action="store_true")
-    p.add_argument("--precision_type", choices=["fp16", "bf16"], default="fp16")
+    p.add_argument("--precision_type", choices=["fp16", "bf16"], default="bf16")
     p.add_argument("--max_lr", type=float, default=2e-3)
     p.add_argument("--weight_decay", type=float, default=2e-4)
     args = p.parse_args()
